feat: divide question, modify third question
